2x Scaling

+++ types.h
@@ -589,7 +589,7 @@ typedef struct gb_state {
     bool lcd_entered_hblank; // Set at the end of every HBlank.
     bool lcd_entered_vblank; // Set at the beginning of every VBlank.
 
-    u32* TextureBuffer;
+    u16* TextureBuffer;
     u16* NextBuffer;
 
     bool isSRAMDisabled; //Writing 0 to MBC 1 turns off access to external RAM
